refactor: parseVerified
